requirements.txt
torch
librosa
pydub
jiwer
PyYAML
pytest